First Animation !
MainView.axaml
Add a Border to add a transition property to InfoBar.

Styles.axaml
Add a class for InfoBar and default values.
Add a class for un/collapsed InfoBar.

StatusBarViewModel.cs
Add a property for un/collapsed InfoBar.
Add a delay before (auto)hiding the InfoBar with isVisible so the animation can be seen.